+++ src/gwt/src/org/rstudio/studio/client/common/filetypes/TextFileType.java
@@ -347,6 +347,7 @@ public class TextFileType extends EditableFileType
       {
          results.add(commands.editRmdFormatOptions());
          results.add(commands.knitWithParameters());
+         results.add(commands.clearKnitrCache());
          results.add(commands.restartRClearOutput());
          results.add(commands.restartRRunAllChunks());
          results.add(commands.notebookCollapseAllOutput());