+++ NEWS.md
@@ -65,6 +65,7 @@
 * Add R Markdown pane to global options dialog
 * Build tab and preview support for R Markdown websites
 * Various enhancements for authoring with the bookdown package
+* Add command to clear knitr cache
 * Update pandoc to 1.17.2
 * Update MathJax to 2.6.1